requirements.txt
boto3==1.22.8
botocore==1.25.8
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.5
idna==3.3
jmespath==1.0.0
lxml==4.9.1
numpy==1.23.3
pika==1.2.1
portalocker==2.5.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.9.13
requests==2.27.1
s3transfer==0.5.2
sacrebleu==2.2.1
six==1.16.0
tabulate==0.8.10
urllib3==1.26.9
